+++ src/function/index.ts
@@ -1,4 +1,5 @@
 export { default as nodeToLayers } from './nodeToLayers';
 export { default as nodeToGroup } from './nodeToGroup';
-export { default as nodeToSketchSymbol } from './nodeToSketchSymbol';
+export { default as nodeToSymbol } from './nodeToSymbol';
 export { default as adjustSymbolLayout } from './adjustSymbolParams';
+export { setConfig } from './setting';